Update Storage strategies (incomplete)
