+++ src/main/java/io/github/social/nostr/relay/def/IEventService.java
@@ -8,8 +8,6 @@ public interface IEventService {
     static final String DB_ERROR     = "error: Could not connect to database.";
     static final String REG_REQUIRED = "blocked: pubkey must be registered.";
 
-    // public static final IEventService INSTANCE = new EventLocalStorageDataService();
-    // public static final IEventService INSTANCE = new EventCacheDataService();
     public static final IEventService INSTANCE = new EventDocumentDataService();
 
     String checkRegistration(final EventData eventData);
